[Auditorium increased]
