+++ Lodgify.Cinema.Domain/Resources/BusinessMessage.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="InvalidAuditorium" xml:space="preserve">
+    <value>Id {0} invalid for the auditorium, the id must be greater than 0 and less than 4</value>
+  </data>
   <data name="InvalidToken" xml:space="preserve">
     <value>Invalid Token</value>
   </data>
